@@ -22,8 +22,8 @@ namespace Ditto.Bot.Data.API.Rest
         public string Fortune()
         {
             var fortune = Call("v1/fortunes", new[] {
-                new RestSharp.Parameter("limit", "1", RestSharp.ParameterType.GetOrPost),
-                new RestSharp.Parameter("skip", Randomizer.Static.New(0, 543), RestSharp.ParameterType.GetOrPost) // 2018-01-26: limit of 544 results.
+                new Parameter("limit", "1"),
+                new Parameter("skip", Randomizer.Static.New(0, 543)) // 2018-01-26: limit of 544 results.
             });
             return fortune.FirstOrDefault()?.Message ?? string.Empty;
         }
