@@ -67,7 +67,7 @@ namespace Ditto.Bot.Modules.Utility.Linking
                         try
                         {
                             Log.Debug($"[BDO] Posting \"{item.Title}\" ({LinkType.BDO.ToString()}) in \"{link.Guild.Name}:{link.Channel.Name}\"");
-                            await channel.EmbedAsync(ParseEmbed(item.Link, channel.Guild, feed.Channel?.Title),
+                            await channel.EmbedAsync(await ParseEmbed(item.Link, channel.Guild, feed.Channel?.Title).ConfigureAwait(false),
                                 options: new RequestOptions() { RetryMode = RetryMode.AlwaysFail }
                             );
                             listUrls.Add(item.Guid ?? item.Link);
@@ -135,9 +135,9 @@ namespace Ditto.Bot.Modules.Utility.Linking
             }
         }
 
-        protected static EmbedBuilder ParseEmbed(string url, IGuild guild, string channelName)
+        protected static async Task<EmbedBuilder> ParseEmbed(string url, IGuild guild, string channelName)
         {
-            var htmlCode = WebHelper.GetSourceCode(url);
+            var htmlCode = await WebHelper.GetSourceCodeAsync(url).ConfigureAwait(false);
             var metaInfo = WebHelper.GetMetaInfoFromHtml(htmlCode);
             var siteName = metaInfo.FirstOrDefault(e => e.Name.Equals("og:site_name", StringComparison.CurrentCultureIgnoreCase))?.Value;
             var title = metaInfo.FirstOrDefault(e => e.Name.Equals("title", StringComparison.CurrentCultureIgnoreCase))?.Value
